* release/v2.2: (64 commits)
  Updated readme
  Added release binaries
  Updated copyright
  Bump version
  Fixed addon category/action
  Updated addon spec
  Updated read min requirements
  Updated screenshots
  Added duplicate feature
  Added imeoption and hide keyboard functionality
  Tidied layout
  Added basic image description functionality
  Updated dependencies and build tools
  New icon
  Removed icon sketch file
  Send update broadcast on activity resume
  Set base activity
  Added app type meta data for GrowUpdater
  Added custom addon action for GrowUpdater (optional addon)
  Created alpha1 binaries
  ...

+++ app/src/main/java/me/anon/lib/manager/PlantManager.java
@@ -21,6 +21,7 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.experimental.Accessors;
 import me.anon.grow.MainApplication;
+import me.anon.lib.helper.AddonHelper;
 import me.anon.lib.helper.EncryptionHelper;
 import me.anon.lib.helper.GsonHelper;
 import me.anon.lib.task.AsyncCallback;
@@ -58,6 +59,11 @@ public class PlantManager
 
 	public ArrayList<Plant> getSortedPlantList(@Nullable Garden garden)
 	{
+		if (MainApplication.isFailsafe())
+		{
+			return new ArrayList<>();
+		}
+
 		synchronized (this.mPlants)
 		{
 			SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
@@ -91,8 +97,11 @@ public class PlantManager
 	{
 		synchronized (this.mPlants)
 		{
-			this.mPlants.clear();
-			this.mPlants.addAll(plants);
+			if (!MainApplication.isFailsafe())
+			{
+				this.mPlants.clear();
+				this.mPlants.addAll(plants);
+			}
 		}
 	}
 
@@ -100,8 +109,11 @@ public class PlantManager
 	{
 		synchronized (this.mPlants)
 		{
-			mPlants.add(plant);
-			save();
+			if (!MainApplication.isFailsafe())
+			{
+				mPlants.add(plant);
+				save();
+			}
 		}
 	}
 
@@ -109,19 +121,22 @@ public class PlantManager
 	{
 		synchronized (this.mPlants)
 		{
-			// Delete images
-			ArrayList<String> imagePaths = mPlants.get(plantIndex).getImages();
-			for (String filePath : imagePaths)
+			if (!MainApplication.isFailsafe())
 			{
-				new File(filePath).delete();
+				// Delete images
+				ArrayList<String> imagePaths = mPlants.get(plantIndex).getImages();
+				for (String filePath : imagePaths)
+				{
+					new File(filePath).delete();
+				}
+
+				// Remove plant
+				mPlants.remove(plantIndex);
+
+				// Remove from shared prefs
+				SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
+				prefs.edit().remove(String.valueOf(plantIndex)).apply();
 			}
-
-			// Remove plant
-			mPlants.remove(plantIndex);
-
-			// Remove from shared prefs
-			SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
-			prefs.edit().remove(String.valueOf(plantIndex)).apply();
 		}
 	}
 
@@ -140,6 +155,11 @@ public class PlantManager
 
 	public void load()
 	{
+		if (MainApplication.isFailsafe())
+		{
+			return;
+		}
+
 		if (FileManager.getInstance().fileExists(FILES_DIR + "/plants.json"))
 		{
 			String plantData;
@@ -177,7 +197,10 @@ public class PlantManager
 	{
 		synchronized (this.mPlants)
 		{
-			save(null);
+			if (!MainApplication.isFailsafe())
+			{
+				save(null);
+			}
 		}
 	}
 
@@ -185,48 +208,53 @@ public class PlantManager
 	{
 		synchronized (mPlants)
 		{
-			new AsyncTask<Void, Void, Void>()
+			if (!MainApplication.isFailsafe())
 			{
-				@Override protected Void doInBackground(Void... voids)
+				AddonHelper.broadcastPlantList(context);
+
+				new AsyncTask<Void, Void, Void>()
 				{
-					synchronized (mPlants)
+					@Override protected Void doInBackground(Void... voids)
 					{
-						if (MainApplication.isEncrypted())
+						synchronized (mPlants)
 						{
-							if (TextUtils.isEmpty(MainApplication.getKey()))
+							if (MainApplication.isEncrypted())
 							{
-								return null;
+								if (TextUtils.isEmpty(MainApplication.getKey()))
+								{
+									return null;
+								}
+
+								FileManager.getInstance().writeFile(FILES_DIR + "/plants.json", EncryptionHelper.encrypt(MainApplication.getKey(), GsonHelper.parse(mPlants)));
 							}
-
-							FileManager.getInstance().writeFile(FILES_DIR + "/plants.json", EncryptionHelper.encrypt(MainApplication.getKey(), GsonHelper.parse(mPlants)));
+							else
+							{
+								FileManager.getInstance().writeFile(FILES_DIR + "/plants.json", GsonHelper.parse(mPlants));
+							}
 						}
-						else
+
+						return null;
+					}
+
+					@Override protected void onPostExecute(Void aVoid)
+					{
+						if (callback != null)
 						{
-							FileManager.getInstance().writeFile(FILES_DIR + "/plants.json", GsonHelper.parse(mPlants));
+							callback.callback();
+						}
+
+						if (new File(FILES_DIR + "/plants.json").length() == 0 || !new File(FILES_DIR + "/plants.json").exists())
+						{
+							Toast.makeText(context, "There was a fatal problem saving the plant data, please backup this data", Toast.LENGTH_LONG).show();
+							String sendData = GsonHelper.parse(mPlants);
+							Intent share = new Intent(Intent.ACTION_SEND);
+							share.setType("text/plain");
+							share.putExtra(Intent.EXTRA_TEXT, "== WARNING : PLEASE BACK UP THIS DATA == \r\n\r\n " + sendData);
+							context.startActivity(share);
 						}
 					}
-
-					return null;
-				}
-
-				@Override protected void onPostExecute(Void aVoid)
-				{
-					if (callback != null)
-					{
-						callback.callback();
-					}
-
-					if (new File(FILES_DIR + "/plants.json").length() == 0 || !new File(FILES_DIR + "/plants.json").exists())
-					{
-						Toast.makeText(context, "There was a fatal problem saving the plant data, please backup this data", Toast.LENGTH_LONG).show();
-						String sendData = GsonHelper.parse(mPlants);
-						Intent share = new Intent(Intent.ACTION_SEND);
-						share.setType("text/plain");
-						share.putExtra(Intent.EXTRA_TEXT, "== WARNING : PLEASE BACK UP THIS DATA == \r\n\r\n " + sendData);
-						context.startActivity(share);
-					}
-				}
-			}.executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
+				}.executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
+			}
 		}
 	}
 }
